Day 7, full card sorting
Returns the cards sorted by the full value, ie `2AAAA` beats `3333A` as it is four of a kind on aces.
This isn't quite what the question asks for, as this requires that once we have the kind of hand, we simply sort by the first, second etc. card. In which case `3333A` beats `2AAAA`.